DX-914: Move `which.rb` into the `lib` directory
